refactor file layout
